Msg Sender to clients for game server owners
Well this is my 1st proj on sourceforge, and i'm a beginner. This
project has 2 things.

1 - Releaser - server
2 - Catcher - client

1-----
Releaser is the server program. It contains 3 files.

--settings--
There is a file named "settings.ini". Open it and change the number in
there..It calls the port which your pc allow to make connections through
that port.

--msg--
there is a txt file named "msg". edit it and write your own message.
whenever someone connect you, the server will send the msg.

--releaser--
the server

PS:- Open the port in your firewall before start the server

2-----
Simply, this is the client. type the ip and port that server pc's. "Get
it" button will get the message for you from the server's pc.

I play games, so I thought that this will helpful to the server owners,
so they can send any message to the ones that connects to their server.

I hope to see your comments about my project as this is my first :)

Thanks